finish task session 7

+++ internal/repositories/products_repo.go
@@ -1,20 +1,24 @@
 package repositories
 
 import (
+	"context"
+	"errors"
+	"fmt"
+
 	"github.com/ilhamgepe/prakerja-s7/internal/models"
 	"gorm.io/gorm"
 )
 
 type ProductsRepo interface {
-	GetProducts() (*[]models.Products, error)
+	GetProducts(ctx context.Context) (*[]models.Products, error)
 
-	GetProduct(id uint64) (*models.Products, error)
+	GetProduct(ctx context.Context, id uint64) (*models.Products, error)
 
-	AddProduct(product *models.Products) error
+	AddProduct(ctx context.Context, product *models.Products) error
 
-	UpdateProduct(id uint64, product *models.Products) error
+	UpdateProduct(ctx context.Context, id uint64, product *models.Products) error
 
-	DeleteProduct(id uint64) error
+	DeleteProduct(ctx context.Context, id uint64) error
 }
 
 type productsRepo struct {
@@ -27,7 +31,7 @@ func NewProductsRepo(db *gorm.DB) ProductsRepo {
 	}
 }
 
-func (pr *productsRepo) GetProducts() (products *[]models.Products, err error) {
+func (pr *productsRepo) GetProducts(ctx context.Context) (products *[]models.Products, err error) {
 	err = pr.DB.Find(&products).Error
 	if err != nil {
 		return nil, err
@@ -35,21 +39,41 @@ func (pr *productsRepo) GetProducts() (products *[]models.Products, err error) {
 	return products, nil
 }
 
-func (pr *productsRepo) GetProduct(id uint64) (product *models.Products, err error) {
+func (pr *productsRepo) GetProduct(ctx context.Context, id uint64) (product *models.Products, err error) {
 	err = pr.DB.First(&product, id).Error
 	return
 }
 
-func (pr *productsRepo) AddProduct(p *models.Products) (err error) {
-	return pr.DB.Create(&p).Error
-}
-
-func (pr *productsRepo) UpdateProduct(id uint64, p *models.Products) error {
-	var products *models.Products
-	err := pr.DB.First(&products, id).Error
+func (pr *productsRepo) AddProduct(ctx context.Context, p *models.Products) (err error) {
+	email := ctx.Value("email")
+	var user models.Users
+	err = pr.DB.Where("email = ?", email).First(&user).Error
 	if err != nil {
 		return err
 	}
+
+	p.UserID = user.ID
+	return pr.DB.Create(&p).Error
+}
+
+func (pr *productsRepo) UpdateProduct(ctx context.Context, id uint64, p *models.Products) error {
+	email := ctx.Value("email")
+	var user models.Users
+	err := pr.DB.Where("email = ?", email).First(&user).Error
+	if err != nil {
+		return err
+	}
+
+	p.UserID = user.ID
+	var products *models.Products
+	err = pr.DB.First(&products, id).Error
+	if err != nil {
+		return err
+	}
+	if products.UserID != user.ID {
+		return errors.New("you have no permission to update this product")
+	}
+
 	products.Name = p.Name
 	products.Price = p.Price
 	if err := pr.DB.Save(products).Error; err != nil {
@@ -58,6 +82,22 @@ func (pr *productsRepo) UpdateProduct(id uint64, p *models.Products) error {
 	return nil
 }
 
-func (pr *productsRepo) DeleteProduct(id uint64) error {
+func (pr *productsRepo) DeleteProduct(ctx context.Context, id uint64) error {
+	email := ctx.Value("email")
+	fmt.Println(email)
+	var user models.Users
+	err := pr.DB.Where("email = ?", email).First(&user).Error
+	if err != nil {
+		return err
+	}
+
+	var products *models.Products
+	err = pr.DB.First(&products, id).Error
+	if err != nil {
+		return err
+	}
+	if products.UserID != user.ID {
+		return errors.New("you have no permission to update this product")
+	}
 	return pr.DB.Delete(&models.Products{}, id).Error
 }
